:bug: Fixed issue with vpn device, in infra (#209)
